1.1 : Divide StyleSheet
Divide StyleSheet from Reservation.js
