+++ styles/screens/ReserveStyle.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "white",
+  },
+  pageTitle: {
+    fontSize: 40,
+  },
+});
+
+export default styles;
